# account-service/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Account Microservice API
  description: API for managing accounts in the financial application
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
tags:
  - name: Accounts
    description: Operations related to account management
    
paths:
  /accounts:
    get:
      tags:
        - Accounts
      summary: Get all accounts
      operationId: getAllAccounts
      parameters:
        - name: x-guid
          in: header
          required: true
          schema:
            type: string
            maxLength: 60
            pattern: '^[{]?[0-9a-fA-F]{8}-([0-9a-fA-F]{4}-){3}[0-9a-fA-F]{12}[}]?$'
        - name: x-channel
          in: header
          required: true
          schema:
            type: string
            maxLength: 2
            pattern: '^\d+$'
        - name: x-medium
          in: header
          required: true
          schema:
            type: string
            maxLength: 2
            pattern: '^\d+$'
        - name: x-app
          in: header
          required: true
          schema:
            type: string
            maxLength: 5
            pattern: '^\d+$'
        - name: x-session
          in: header
          required: true
          schema:
            type: string
            maxLength: 100
      responses:
        '200':
          description: A list of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountRequest'
    post:
      tags:
        - Accounts
      summary: Create new account
      operationId: createAccount
      parameters:
        - name: x-guid
          in: header
          required: true
          schema:
            type: string
            maxLength: 60
            pattern: '^[{]?[0-9a-fA-F]{8}-([0-9a-fA-F]{4}-){3}[0-9a-fA-F]{12}[}]?$'
        - name: x-channel
          in: header
          required: true
          schema:
            type: string
            maxLength: 2
            pattern: '^\d+$'
        - name: x-medium
          in: header
          required: true
          schema:
            type: string
            maxLength: 2
            pattern: '^\d+$'
        - name: x-app
          in: header
          required: true
          schema:
            type: string
            maxLength: 5
            pattern: '^\d+$'
        - name: x-session
          in: header
          required: true
          schema:
            type: string
            maxLength: 100
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountRequest'
      responses:
        '201':
          description: Account created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountRequest'
  /accounts/{id}:
    get:
      tags:
        - Accounts
      summary: Get account by ID
      operationId: getAccountById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: x-guid
          in: header
          required: true
          schema:
            type: string
            maxLength: 60
            pattern: '^[{]?[0-9a-fA-F]{8}-([0-9a-fA-F]{4}-){3}[0-9a-fA-F]{12}[}]?$'
        - name: x-channel
          in: header
          required: true
          schema:
            type: string
            maxLength: 2
            pattern: '^\d+$'
        - name: x-medium
          in: header
          required: true
          schema:
            type: string
            maxLength: 2
            pattern: '^\d+$'
        - name: x-app
          in: header
          required: true
          schema:
            type: string
            maxLength: 5
            pattern: '^\d+$'
        - name: x-session
          in: header
          required: true
          schema:
            type: string
            maxLength: 100
      responses:
        '200':
          description: Account found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountRequest'
        '404':
          description: Account not found
    put:
      tags:
        - Accounts
      summary: Update an existing account
      operationId: updateAccount
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: x-guid
          in: header
          required: true
          schema:
            type: string
            maxLength: 60
            pattern: '^[{]?[0-9a-fA-F]{8}-([0-9a-fA-F]{4}-){3}[0-9a-fA-F]{12}[}]?$'
        - name: x-channel
          in: header
          required: true
          schema:
            type: string
            maxLength: 2
            pattern: '^\d+$'
        - name: x-medium
          in: header
          required: true
          schema:
            type: string
            maxLength: 2
            pattern: '^\d+$'
        - name: x-app
          in: header
          required: true
          schema:
            type: string
            maxLength: 5
            pattern: '^\d+$'
        - name: x-session
          in: header
          required: true
          schema:
            type: string
            maxLength: 100
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountRequest'
      responses:
        '200':
          description: Account updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountRequest'
        '404':
          description: Account not found
    delete:
      tags:
        - Accounts
      summary: Delete account by ID
      operationId: deleteAccount
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: x-guid
          in: header
          required: true
          schema:
            type: string
            maxLength: 60
            pattern: '^[{]?[0-9a-fA-F]{8}-([0-9a-fA-F]{4}-){3}[0-9a-fA-F]{12}[}]?$'
        - name: x-channel
          in: header
          required: true
          schema:
            type: string
            maxLength: 2
            pattern: '^\d+$'
        - name: x-medium
          in: header
          required: true
          schema:
            type: string
            maxLength: 2
            pattern: '^\d+$'
        - name: x-app
          in: header
          required: true
          schema:
            type: string
            maxLength: 5
            pattern: '^\d+$'
        - name: x-session
          in: header
          required: true
          schema:
            type: string
            maxLength: 100
      responses:
        '204':
          description: Account deleted successfully
        '404':
          description: Account not found
  /accounts/filter/{customerId}:
    get:
      tags:
        - Accounts
      summary: Get all accounts by customerId
      operationId: getAccountsByCustomerId
      parameters:        
        - name: customerId
          in: path
          required: true
          schema:
            type: integer        
        - name: x-guid
          in: header
          required: true
          schema:
            type: string
            maxLength: 60
            pattern: '^[{]?[0-9a-fA-F]{8}-([0-9a-fA-F]{4}-){3}[0-9a-fA-F]{12}[}]?$'
        - name: x-channel
          in: header
          required: true
          schema:
            type: string
            maxLength: 2
            pattern: '^\d+$'
        - name: x-medium
          in: header
          required: true
          schema:
            type: string
            maxLength: 2
            pattern: '^\d+$'
        - name: x-app
          in: header
          required: true
          schema:
            type: string
            maxLength: 5
            pattern: '^\d+$'
        - name: x-session
          in: header
          required: true
          schema:
            type: string
            maxLength: 100
      responses:
        '200':
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountRequest'
  /accounts/filter/movement/{accountNumber}:
    get:
      tags:
        - Accounts
      summary: Get all accounts by accountNumber
      operationId: getAccountsByAccountNumber
      parameters:        
        - name: accountNumber
          in: path
          required: true
          schema:
            type: string        
        - name: x-guid
          in: header
          required: true
          schema:
            type: string
            maxLength: 60
            pattern: '^[{]?[0-9a-fA-F]{8}-([0-9a-fA-F]{4}-){3}[0-9a-fA-F]{12}[}]?$'
        - name: x-channel
          in: header
          required: true
          schema:
            type: string
            maxLength: 2
            pattern: '^\d+$'
        - name: x-medium
          in: header
          required: true
          schema:
            type: string
            maxLength: 2
            pattern: '^\d+$'
        - name: x-app
          in: header
          required: true
          schema:
            type: string
            maxLength: 5
            pattern: '^\d+$'
        - name: x-session
          in: header
          required: true
          schema:
            type: string
            maxLength: 100
      responses:
        '200':
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountRequest'                  
components:
  schemas:
    AccountRequest:
      type: object
      properties:
        id:
          type: integer
        accountNumber:
          type: string
        accountType:
          type: string
          enum: [SAVINGS, CURRENT]
        initialBalance:
          type: number
        status:
          type: boolean
        customerId:
          type: integer
      required: [accountNumber, accountType, initialBalance, status, customerId]