# front-end/spf-mfa-flujo-bancario-bp-transaccion/src/app/interfaces/account-request.interface.ts
import { Customer } from "./customer.interface";
export interface AccountRequest {
    id?: number;
    accountNumber: string;
    accountType: 'SAVINGS' | 'CURRENT';
    initialBalance: number;
    status: boolean;
    customerId: number;
    customer?: Customer;
}