# front-end/spf-mfa-flujo-bancario-bp-transaccion/src/app/interfaces/account-response.interface.ts
import { Customer } from "./customer.interface";
export interface AccountRequest {
    accountNumber: string;
    accountType: 'SAVINGS' | 'CURRENT';
    initialBalance: number;
    status: boolean;
    customerId: string;
    customer: Customer;
}